Api details page added back in for GMG, and tested with GG + GMG.
