+++ projects/ui/src/Components/ApiDetails/gloo-gateway-components/SchemaTab/ApiSchemaDisplay.style.tsx
